(GH-2) Optionally do not validate HTTPS certificates
Previously HTTPS requests could fail if they could not be validated (e.g.
missing CA certificates) However on Windows, ruby doesn't use the updated
certificate store which can cause false negatives.  This commit updates the
retrieval method to not validate HTTPS certificates if the SSL_CERT_FILE is not
set.  While a little dangerous, this risk of a man-in-the-middle attack for the
forge API is low.