(GH-2) Use static CA Cert file
Previously the Forge Spec Searcher would turn off HTTPS CA verification if the
SSL_CERT_FILE environment variable was not set, however this is less than ideal.
This commit changes the searcher to instead use a static CA file so we can
still verify the CA certificates.  This commit also adds a Rake task so the
update of the static data can be easily performed.