Support version ranges when resolving dependencies
This change adds support for using version ranges for module
definitions. Currently, module definitions are restricted to using a
specific semantic version, or using versioning shorthand, as the
resolver sets the version to `=#{mod.version}`. If a module has a
version range, this results in an unparsable version (e.g. `=>=1.0.0 <
3.0.0`).