Support using commits for GClone spec searcher
This fixes a bug with the `GClone` spec searcher that caused `git clone`
to fail when provided a commit to clone. Because `git clone` does
not support cloning a specific commit (using a SHA1), it's necessary to
clone the entire repository. This updates the `git clone` command used
by the spec searcher accordingly.