Clone bare repository, fall back to cloning full repository
This updates the `GClone` spec searcher to start by attempting to clone
a bare repository. If cloning a bare repository fails, the spec searcher
falls back to attempting to clone the full repository. Because
Puppetfile supports using commits/SHA1s as refs when installing git
modules, puppetfile-resolver needs to support them as well. Cloning bare
repositories with a commit/SHA1 fails, so falling back to a full clone
ensures that the resolver supports this behavior.

The spec searcher could always clone a full repository instead of
first attempting to clone a bare repository, but this might dramatically
increase the time it takes to clone a module.